Completely revamp server to support full-fledged 9P, SELinux
Utterly eviscerate and rewrite the server to support a subset of
9P as a file service protocol. Serve all RPCs at /srv/n, where n
is a monotonically increasing identifier which identifies the
present connection.

Copy extended attributes (that means SELinux information) into
the rpcfs at /srv/n, and additionally drop permissions to
a properly authenticated user. For now, authentication travels
through PAM (/etc/shadow) on the local machine, and a user
must present a 9P authentication token (afid) in order to
attach to the tree. Once this occurs, drop permissions, which
results in enforcement of both MAC and DAC policy.

Limitations: for now, only allow one connection in per server
instance to minimize complications surrounding asyncio's incompatibility
with os.fork() and friends. This isn't a game-breaker, but will
make further work a lot more complicated. Also, don't write in
full 9P-esque compatibility for server-dictated auth via the
afid, instead fake this with synthetic afids.

For more information about 9P, see the following:
http://man.cat-v.org/plan_9/5/

Signed-off-by: Jay Lang <jaytlang@mit.edu>