Modify the 9P implementation towards RPC-centric append operations
Scrap read, as suggested by Ravi, and instead make append a universal,
one-size-fits-all write-then-read operation. It now takes in data to
be written along with a fid to write to, and returns data in response.
This can be used to retrofit a file-like read operation if no data is
written, but more importantly, it can be used to call an RPC with a
single network traversal instead of the write/read approach's 2.

In order to allow discovery-based navigation as before, modify the Stat
structure to contain a list of children. Make this empty when the system
is a file or for an empty directory, otherwise list out children there.

Additionally, remove length fields for all 9P calls since we don't need
them. The message framework handles length implicitly, and we're not
working in C so don't need buffer sizes.

Nits: change the testing username to something more creative, since I
installed a new machine (:

Signed-off-by: Jay Lang <jaytlang@mit.edu>