Permission root directory, incorporate permissions into 9P stat
Previously, 9P stat forcefully failed instead of properly returning
an error. Fix this by returning Error.EOPENRDF, as if we're reading
the stat out and fail to do so.

The root directory wasn't permissioned before. This would allow an
antagonist to first-level walk an RPC at the top level. Fix this.

Signed-off-by: Jay Lang <jaytlang@mit.edu>